[AIP-21] Move operators.hive_to_samba_operator providers.apache.hive.operators.hive_to_samba
